# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ src/main/java/com/exadel/training/service/impl/ApproveActionServiceImpl.java
@@ -28,7 +28,7 @@ public class ApproveActionServiceImpl implements ApproveActionService {
 
     @Override
     @Transactional
-    public Integer getActionNumber() {
+    public Long getActionNumber() {
         return approveActionDAO.getApproveActionNumber();
     }
